:truck: Move files to a better directory
