+++ types/utils/exceptions.ts
